refactor: trying to reduce duplication
