+++ src/App.jsx
@@ -1,10 +1,8 @@
 import React from 'react';
 import GithubRelease from './components/GithubRelease';
-import GithubPreRelease from './components/GithubPreRelease'
 
 const App = () => (
   <div className="app">
-    <GithubPreRelease />
     <GithubRelease />
   </div>
 );
