+++ src/components/GithubRelease.jsx
@@ -1,36 +1,48 @@
-import {connect} from 'react-redux';
+import { connect } from 'react-redux';
 import React from 'react';
-import {bindActionCreators} from 'redux';
-import { fetchReleases, isCollapsed } from '../actions/githubActions';
-import LoadingComponent from './LoadingComponent';
+import { bindActionCreators } from 'redux';
+import { fetchReleases } from '../actions/githubActions';
+import { isCollapsed } from '../actions/collapsedActions';
 import Markdown from 'react-remarkable';
 import { CollapsibleComponent, CollapsibleHead, CollapsibleContent } from 'react-collapsible-component';
- 
 
-const latestReleaseData = (projects, isCollapsed) => 
+const releaseComponent = (repoName, releaseDescription, className, collapsed) => {
+    return <div className={className}>
+        <div onClick={() => isCollapsed(className, !collapsed)}>
+            <CollapsibleComponent>
+                <CollapsibleHead>{repoName}</CollapsibleHead>
+                <CollapsibleContent isExpanded={collapsed || false}>
+                    <Markdown key={repoName}>{releaseDescription}</Markdown>
+                </CollapsibleContent>
+            </CollapsibleComponent>
+        </div>
+    </div>
+}
+
+const preReleaseData = projects => 
     Object.values(projects.totalReleases).map((repoDetails) => {
-        const latestRelease = Object.values(repoDetails.releases).find(release => {
-            return release.isPrerelease === false
-        });
+        const preRelease = Object.values(repoDetails.releases).find(release => release.isPrerelease === true);
+        const className = `pre-${repoDetails.name}`;
 
+        if(preRelease) {
+            return releaseComponent(repoDetails.name, preRelease.description, className, projects[className]);
+        } else {
+            return  <div><CollapsibleHead>No upcoming relases planned for {repoDetails.name}</CollapsibleHead></div>
+        }
+    })
+
+const latestReleaseData = projects => 
+    Object.values(projects.totalReleases).map((repoDetails) => {
+        const latestRelease = Object.values(repoDetails.releases).find(release => release.isPrerelease === false);
         const className = `latest-${repoDetails.name}`;
 
         if(latestRelease) {
-            return <div className={className}>
-
-                <div onClick={() => isCollapsed(className, !projects[className])}>
-                    <CollapsibleComponent>
-                        <CollapsibleHead>{repoDetails.name}</CollapsibleHead>
-                        <CollapsibleContent isExpanded={projects[className] || false}>
-                            <Markdown key={repoDetails.name}>{latestRelease.description}</Markdown>
-                        </CollapsibleContent>
-                    </CollapsibleComponent>
-                </div>
-            </div>
+            return releaseComponent(repoDetails.name, latestRelease.description, className, projects[className]);
         } else {
-            return  <div><CollapsibleHead>No Latest relases planned for {repoDetails.name}</CollapsibleHead></div>
+            return  <div><CollapsibleHead>No latest relase documented for {repoDetails.name}</CollapsibleHead></div>
         }
-})
+    })
+    
 
 class GithubRelease extends React.Component {
   componentWillMount() {
@@ -38,10 +50,12 @@ class GithubRelease extends React.Component {
   }
 
   render() {
-    if (!this.props.projects.totalReleases) return <LoadingComponent/>;
+    if (!this.props.projects.totalReleases) return <div/>;
     
     return(
         <div>
+            <h1>Upcoming releases</h1>
+            {preReleaseData(this.props.projects, this.props.isCollapsed)}
             <h1>Latest releases</h1>
             {latestReleaseData(this.props.projects, this.props.isCollapsed)}
         </div>
@@ -50,7 +64,10 @@ class GithubRelease extends React.Component {
 }
 
 function mapStateToProps(state) {
-  return {projects: state.github};
+  return {
+      projects: state.github,
+      collapsed: state.collapsed,
+    };
 }
 
 function mapDispatchToProps(dispatch) {
